pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comp-photonics-workflows"
version = "0.1.0"
description = "Closed-loop computational photonics workflow demonstration"
authors = [
    { name = "Generated by ChatGPT", email = "noreply@example.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "scipy",
    "pandas",
    "pyyaml",
    "matplotlib",
    "numba",
    "pytest",
]

[project.scripts]
cpw = "cpw.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
